Refactor PBKDF2 component structure and introduce PBKDF2 feature module.
