+++ components/org.wso2.carbon.identity.hash.provider.pbkdf2/src/test/resources/testng.xml
@@ -1,0 +1,28 @@
+<!--
+  ~ Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="org.wso2.carbon.identity.hash.provider.pbkdf2">
+    <test name="org.wso2.carbon.identity.hash.provider.pbkdf2" preserve-order="false" parallel="false">
+        <classes>
+            <class name="org.wso2.carbon.identity.hash.provider.pbkdf2.PBKDF2HashProviderTest"/>
+            <class name="org.wso2.carbon.identity.hash.provider.pbkdf2.PBKDF2HashProviderFactoryTest"/>
+        </classes>
+    </test>
+</suite>
